Modify translation.csv field structure, rename to normalized.csv
This creates a copy of translation.csv with a few differences in the
fields. Since I couldn't figure out the exact significance of the last
two fields in the original, I removed them.

The second column, instead of being (by default) a copy of the japanese
text, is now empty if the line is untranslated.

There's now a column with "True" or "False" (python style) for whether
the line is translated, making filtering based on that criteria easier.

Adds the 'Speaker' field produced by the changes in translate.py.