Add flask webapp for managing translation CSV
Manually editing the CSV leaves a lot to be desired, in my opinion. This
commit introduces a webapp that allows for simultaneous, live editing of
the translation CSV.

The home page is organized by chapter, and then by subchapter, each
corresponding to a subsection file (i.e. 02_abend_03).

It's not pretty but I've used it to some effect already, and if I had to
guess it's more user friendly than emailing the CSV.